added updateServiceWorker script to separate sw update
