+++ app.js
@@ -2,9 +2,7 @@ import { Clock, NewYearClock } from "./js/clock.js";
 import { waitForAnimation } from "./js/appfunctions.js";
 import { errorHandler } from "./js/error.js";
 import { setInnerHtmlForNotNull } from "./js/functions.js";
-import { registerSW } from 'virtual:pwa-register'
-import Toastify from 'toastify-js'
-import "toastify-js/src/toastify.css"
+// import registerUpdateServiceWorker from "./js/serviceWorkerUpdate"
 
 // DOM nodes
 const animatedCountDuration = 800;
@@ -58,38 +56,7 @@ try {
     errorHandler("Error in clock");
     console.log(error);
 }
-//add service worker update functionality
-//service worker update and offline functionality
-const updateSW = registerSW({
-    onNeedRefresh() {
-        console.log('Update sw, now available, devEnv');
-      Toastify({
-        text: `
-        <h4>A newer version of this page is available!</h4>
-               <br>
-               <a class='do-sw-update'>Click this banner to update and reload</a>
-               `,
-        escapeMarkup: false,
-        offset: {
-          x: 50, // horizontal axis - can be a number or a string indicating unity. eg: '2em'
-          y: -150 // vertical axis - can be a number or a string indicating unity. eg: '2em'
-        },
-        // selector: 'testToast',
-        className:'updateToastify',
-        close: true,
-        gravity: "bottom",
-        duration: -1,
-        onClick() {
-          console.log('Clicking on update to refresh the new service worker');
-          updateSW(true);
-        }
-      }).showToast();
 
-    },
-    onOfflineReady() {
-        console.log('App is offline now');
-    },
-  })
 /*
 if ("serviceWorker" in navigator) {
     // && !/localhost/.test(window.location)) {
