+++ html/about.html
@@ -145,6 +145,7 @@
     <script src="/js/error.js" type="module"></script>
     <script src="/js/loadCustomUI.js" type="module"></script>
     <script src="/js/sidebar.js" async type="module"></script>
+    <script  src="/js/serviceWorkerUpdate.js" type= "module"></script>
     
 </body>
