+++ html/countdown-list.html
@@ -125,6 +125,7 @@
     <script src="/js/loadCustomUI.js" type="module"></script>
     <script src="/js/formupdate.js" type="module" async></script>
     <script src="/js/displayCountdowns.js" type="module"></script>
+    <script  src="/js/serviceWorkerUpdate.js" type= "module"></script>
 
 </body>
