+++ html/today.html
@@ -112,5 +112,6 @@
     <script src="/js/sidebar.js" type="module" async></script>
     <script src="/js/loadCustomUI.js" type="module"></script>
     <script src="/js/today.js" type ="module"></script>
+    <script  src="/js/serviceWorkerUpdate.js" type= "module"></script>
 </body>
 </html>
