+++ index.html
@@ -125,6 +125,7 @@
     <script src="/js/sidebar.js" type="module" async></script>
     <script src="/js/loadCustomUI.js" type="module"></script>
     <script  src="/js/form.js" type= "module" async></script>
+    <script  src="/js/serviceWorkerUpdate.js" type= "module"></script>
         
 </body>
 </html>
